sql_virtual_rag_agent/requirements.txt
langchain
pypdf
python-dotenv
pandas
nltk
spacy
langchain-core
langchain-community
sqlalchemy
requests
sqlglot
ipykernel